Factor out transformData into separate file and update unit test
