+++ components/Tree/index.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react'
 import { Link } from '../../lib/routes'
 import { css } from 'glamor'
-import { ascending, descending, max } from 'd3-array'
+import { max } from 'd3-array'
 import { schemeCategory10 } from 'd3-scale'
 import { color as d3Color } from 'd3-color'
 import CheckIcon from 'react-icons/lib/md/check'
+import { transformData } from './transformData'
 import withT from '../../lib/withT'
 import { swissTime } from '../../lib/utils/format'
 
@@ -101,7 +102,7 @@ class Tree extends Component {
   }
 
   componentWillReceiveProps (nextProps) {
-    this.setState(() => this.transformData(nextProps))
+    this.setState(() => transformData(nextProps))
   }
 
   componentDidUpdate () {
@@ -115,73 +116,6 @@ class Tree extends Component {
 
   componentDidMount () {
     window.addEventListener('resize', this.measure)
-  }
-
-  transformData (props) {
-    let commits = props.commits
-      .map(commit => {
-        return {
-          id: commit.id,
-          date: commit.date,
-          author: commit.author,
-          message: commit.message,
-          parentIds: commit.parentIds,
-          milestones: props.milestones.filter(o => {
-            return o.commit.id === commit.id
-          })
-        }
-      })
-      .sort(function (a, b) {
-        return ascending(new Date(a.date), new Date(b.date))
-      })
-
-    let parentNodes = new Map()
-    let links = []
-
-    commits.forEach(commit => {
-      commit.setListItemRef = ref => {
-        commit.listItemRef = ref
-      }
-      commit.setNodeRef = ref => {
-        commit.nodeRef = ref
-      }
-      commit.setMilestoneBarRef = ref => {
-        commit.milestoneBarRef = ref
-      }
-      commit.data = {
-        slotIndex: null
-      }
-      if (!commit.parentIds) return
-      commit.parentIds.forEach(parentId => {
-        let children = []
-        if (parentNodes.has(parentId)) {
-          children = [...parentNodes.get(parentId)]
-        }
-        if (!children.indexOf(commit.id) === -1) {
-          return
-        }
-        children.push(commit.id)
-        parentNodes.set(parentId, children)
-        links.push({
-          sourceId: parentId,
-          destinationId: commit.id
-        })
-      })
-    })
-
-    links.forEach(link => {
-      link.setRef = ref => {
-        link.ref = ref
-      }
-    })
-
-    assignSlots(commits, parentNodes)
-
-    return {
-      commits: commits,
-      links: links,
-      parentNodes: parentNodes
-    }
   }
 
   measure () {
@@ -401,53 +335,3 @@ class Tree extends Component {
 }
 
 export default withT(Tree)
-
-const getPaths = (commits, parentNodes) => {
-  // Walks and collects all possible upward paths on the tree.
-  let paths = []
-  let pathsToWalk = [[commits[0].id]]
-
-  do {
-    let path = pathsToWalk[0]
-    let nextId = path.pop()
-    do {
-      let children = parentNodes.get(nextId)
-      path.push(nextId)
-      nextId = null
-      if (children && !!children.length) {
-        nextId = children.pop()
-        children.forEach(child => {
-          pathsToWalk.push([...path, child])
-        })
-      }
-    } while (nextId)
-    paths.push(path)
-    pathsToWalk.shift()
-  } while (pathsToWalk.length)
-
-  return paths
-}
-
-const getOrderedPaths = paths => {
-  // TODO: More sophisticated ordering.
-  return paths.sort(function (a, b) {
-    return descending(a.length, b.length)
-  })
-}
-
-const assignSlots = (commits, parentNodes) => {
-  let paths = getPaths(commits, parentNodes)
-  let orderedPaths = getOrderedPaths(paths)
-  commits.sort(function (a, b) {
-    return descending(new Date(a.date), new Date(b.date))
-  })
-
-  commits.forEach(commit => {
-    orderedPaths.some((orderedPath, i) => {
-      if (orderedPath.indexOf(commit.id) > -1) {
-        commit.data.slotIndex = orderedPaths.length - 1 - i
-        return true
-      }
-    })
-  })
-}
